Moved the printer to a shared folder
