+++ src/errorHandler.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const printHelper = require('./helper/print');
+const printer = require('./shared/printer');
 
 const TemplateNotFoundError = require('./errors/templateNotFoundError');
 const TemplateRenderingError = require('./errors/templateRenderingError');
@@ -8,36 +8,36 @@ const TemplateInvalidError = require('./errors/templateInvalidError');
 
 module.exports = (isVerbose, err) => {
     if (err instanceof TemplateNotFoundError) {
-		printHelper.error('Template could not be found');
+		printer.error('Template could not be found');
 
 		if (isVerbose) {
-			printHelper.error(err.innerError.message);
+			printer.error(err.innerError.message);
 		}
 
-		printHelper.info(`Try installing it with npm install -g ${err.templateName}`);
+		printer.info(`Try installing it with npm install -g ${err.templateName}`);
 
 	} else if(err instanceof TemplateRenderingError) {
-		printHelper.error('Template could not be rendered');
+		printer.error('Template could not be rendered');
 
 		if (isVerbose) {
-			printHelper.error(err.innerError.message);
+			printer.error(err.innerError.message);
 		}
 
-		printHelper.info('The file has been created but left empty');
+		printer.info('The file has been created but left empty');
 
 	} else if (err instanceof TemplateInvalidError) {
-		printHelper.error('Template found but is invalid!');
+		printer.error('Template found but is invalid!');
 
 		if (isVerbose) {
-			printHelper.info(`The template ${err.template} does not expose a valid template function`);
-			printHelper.info(`If you are the author of the template then please check the documentation at https://github.com/dejanfajfar/tttemplate`)
+			printer.info(`The template ${err.template} does not expose a valid template function`);
+			printer.info(`If you are the author of the template then please check the documentation at https://github.com/dejanfajfar/tttemplate`)
 		}
 	} else {
-		printHelper.error('There was a catastrophic error!');
+		printer.error('There was a catastrophic error!');
 
 		if (isVerbose) {
-			printHelper.error(err);
-			printHelper.error(err.stack);
+			printer.error(err);
+			printer.error(err.stack);
 		}
 	}
